Factor out WidgetArena into separate mod.
